+++ lib/groupitem_metadata_providers/groupitem_metadata_providers.dart
@@ -5,13 +5,14 @@ import 'dart:async' as async;
 
 import 'package:bwu_datagrid/bwu_datagrid.dart';
 import 'package:bwu_datagrid/datagrid/helpers.dart';
-import 'package:bwu_datagrid/formatters/formatters.dart';
-import 'package:bwu_datagrid/editors/editors.dart';
-import 'package:bwu_datagrid/core/core.dart';
+import 'package:bwu_datagrid/formatters/formatters.dart' as fm;
+import 'package:bwu_datagrid/editors/editors.dart' as ed;
+import 'package:bwu_datagrid/core/core.dart' as core;
 import 'package:bwu_datagrid/dataview/dataview.dart';
+import 'package:bwu_datagrid/plugins/plugin.dart';
 
 
-class DefaultGroupCellFormatter extends Formatter {
+class DefaultGroupCellFormatter extends fm.Formatter {
 
   GroupItemMetadataProvider giMetadataProvider;
 
@@ -20,7 +21,8 @@ class DefaultGroupCellFormatter extends Formatter {
    @override
    void call(dom.HtmlElement target, int row, int cell, String value, Column columnDef, /*Item/Map*/ item) {
     if (!giMetadataProvider.enableExpandCollapse) {
-      return item['title'];
+      target.append(item.title);
+      return;
     }
 
     var indentation = '${item.level * 15}px';
@@ -30,24 +32,26 @@ class DefaultGroupCellFormatter extends Formatter {
         ..append(
             new dom.SpanElement()
                 ..classes.add('${giMetadataProvider.toggleCssClass}')
-                ..classes.add('${item['collapsed'] ? giMetadataProvider.toggleCollapsedCssClass : giMetadataProvider.toggleExpandedCssClass}')
+                ..classes.add('${item.isCollapsed ? giMetadataProvider.toggleCollapsedCssClass : giMetadataProvider.toggleExpandedCssClass}')
                 ..style.marginLeft = '${indentation}px')
-        ..append(new dom.SpanElement()
+        ..append(
+            new dom.SpanElement()
                 ..classes.add('${giMetadataProvider.groupTitleCssClass}')
                 ..attributes['level']='${item.level}'
-                ..text = item.title);
+                ..append(item.title));
   }
 }
 
-class DefaultTotalsCellFormatter extends Formatter {
+class DefaultTotalsCellFormatter extends core.GroupTotalsFormatter {
   GroupItemMetadataProvider giMetadataProvider;
 
   @override
-  void call(dom.HtmlElement target, int row, int cell, String value, Column columnDef, /*Item/Map*/ item) {
+  void call(dom.HtmlElement target, core.GroupTotals totals, Column columnDef) {
     if(columnDef.groupTotalsFormatter != null) {
-      columnDef.groupTotalsFormatter(target, row, cell, null, columnDef, item);
+      columnDef.groupTotalsFormatter(target, totals, columnDef);
+    } else {
+      target.innerHtml = '';
     }
-    target.innerHtml = '';
   }
 }
 
@@ -60,38 +64,48 @@ class DefaultTotalsCellFormatter extends Formatter {
  * If "grid.registerPlugin(...)" is not called, expand & collapse will not work.
  *
  */
-class GroupItemMetadataProvider {
-  BwuDatagrid _grid;
+class GroupItemMetadataProvider extends Plugin {
+  //BwuDatagrid _grid;
 
-  String groupCssClass= "bwu-datagrid-group";
-  String groupTitleCssClass= "bwu-datagrid-group-title";
-  String totalsCssClass= "bwu-datagrid-group-totals";
-  bool groupFocusable= true;
-  bool totalsFocusable= false;
-  String toggleCssClass= "bwu-datagrid-group-toggle";
-  String toggleExpandedCssClass= "expanded";
-  String toggleCollapsedCssClass= "collapsed";
+  String groupCssClass;
+  String groupTitleCssClass;
+  String totalsCssClass;
+  bool groupFocusable;
+  bool totalsFocusable;
+  String toggleCssClass;
+  String toggleExpandedCssClass;
+  String toggleCollapsedCssClass;
   bool enableExpandCollapse= true;
-  Formatter _groupFormatter;
-  Formatter _totalsFormatter;
+  fm.Formatter _groupFormatter;
+  core.GroupTotalsFormatter _totalsFormatter;
 
-  GroupItemMetadataProvider({this.groupCssClass, this.groupTitleCssClass, this.totalsCssClass,
-      this.groupFocusable, this.totalsFocusable, this.toggleCssClass, this.toggleExpandedCssClass,
-      this.toggleCollapsedCssClass, this.enableExpandCollapse, Formatter groupFormatter, Formatter totalsFormatter}) {
-    _groupFormatter = groupFormatter;
-    _totalsFormatter = totalsFormatter;
-    _gridClickSubscription = _grid.onBwuClick.listen(_handleGridClick);
-    _gridKeyDownSubscription = _grid.onBwuKeyDown.listen(_handleGridKeyDown);
+  GroupItemMetadataProvider({this.groupCssClass: "bwu-datagrid-group",
+    this.groupTitleCssClass: "bwu-datagrid-group-title",
+        this.totalsCssClass: "bwu-datagrid-group-totals",
+      this.groupFocusable: true, this.totalsFocusable: false,
+          this.toggleCssClass: "bwu-datagrid-group-toggle",
+              this.toggleExpandedCssClass: "expanded",
+      this.toggleCollapsedCssClass: "collapsed", this.enableExpandCollapse: true,
+          fm.Formatter groupFormatter, core.GroupTotalsFormatter totalsFormatter}) {
+    _groupFormatter = groupFormatter = new DefaultGroupCellFormatter(this);
+    _totalsFormatter = totalsFormatter = new DefaultTotalsCellFormatter();
   }
 
-  Formatter _getGroupFormatter() {
+  @override
+  void init(BwuDatagrid grid) {
+    super.init(grid);
+    _gridClickSubscription = grid.onBwuClick.listen(_handleGridClick);
+    _gridKeyDownSubscription = grid.onBwuKeyDown.listen(_handleGridKeyDown);
+  }
+
+  fm.Formatter _getGroupFormatter() {
     if(_groupFormatter != null) {
       return _groupFormatter;
     }
     return new DefaultGroupCellFormatter(this);
   }
 
-  Formatter _getTotalsFormatter() {
+  core.GroupTotalsFormatter _getTotalsFormatter() {
     if(_totalsFormatter != null) {
       return _totalsFormatter;
     }
@@ -101,6 +115,7 @@ class GroupItemMetadataProvider {
   async.StreamSubscription _gridClickSubscription;
   async.StreamSubscription _gridKeyDownSubscription;
 
+  @override
   void destroy() {
     if(_gridClickSubscription != null) {
       _gridClickSubscription.cancel();
@@ -110,11 +125,11 @@ class GroupItemMetadataProvider {
     }
   }
 
-  void _handleGridClick(Click e) {
+  void _handleGridClick(core.Click e) {
     BwuDatagrid grid = e.sender;
-    ItemBase item = grid.getDataItem(e.cell.row);
-    if (item != null && item is Group && (e.causedBy.target as dom.HtmlElement).classes.contains(toggleCssClass)) {
-      var range = _grid.getRenderedRange();
+    core.ItemBase item = grid.getDataItem(e.cell.row);
+    if (item != null && item is core.Group && (e.causedBy.target as dom.HtmlElement).classes.contains(toggleCssClass)) {
+      var range = grid.getRenderedRange();
       if(grid.dataProvider is DataView) {
         var dp = grid.dataProvider as DataView;
         dp.setRefreshHints({
@@ -124,9 +139,9 @@ class GroupItemMetadataProvider {
 
 
         if (item.isCollapsed) {
-          dp.expandGroup(item.groupingKey);
+          dp.expandGroup([item.groupingKey]);
         } else {
-          dp.collapseGroup(item.groupingKey);
+          dp.collapseGroup([item.groupingKey]);
         }
       }
       e.stopImmediatePropagation();
@@ -135,14 +150,14 @@ class GroupItemMetadataProvider {
   }
 
   // TODO:  add -/+ handling
-  void _handleGridKeyDown(KeyDown e) {
+  void _handleGridKeyDown(core.KeyDown e) {
     BwuDatagrid grid = e.sender;
     if (enableExpandCollapse && (e.causedBy.which == dom.KeyCode.SPACE)) {
       var activeCell = grid.getActiveCell();
       if (activeCell != null) {
-        ItemBase item = grid.getDataItem(activeCell.row);
-        if (item != null && item is Group) {
-          var range = _grid.getRenderedRange();
+        core.ItemBase item = grid.getDataItem(activeCell.row);
+        if (item != null && item is core.Group) {
+          var range = grid.getRenderedRange();
 
           if(grid.dataProvider is DataView) {
             var dp = grid.dataProvider as DataView;
@@ -153,9 +168,9 @@ class GroupItemMetadataProvider {
             });
 
             if (item.isCollapsed) {
-              dp.expandGroup(item.groupingKey);
+              dp.expandGroup([item.groupingKey]);
             } else {
-              dp.collapseGroup(item.groupingKey);
+              dp.collapseGroup([item.groupingKey]);
             }
           }
 
@@ -166,7 +181,7 @@ class GroupItemMetadataProvider {
     }
   }
 
-  RowMetadata getGroupRowMetadata(ItemBase item) {
+  RowMetadata getGroupRowMetadata(core.ItemBase item) {
     return new RowMetadata(
       selectable: false,
       focusable: groupFocusable,
@@ -180,7 +195,7 @@ class GroupItemMetadataProvider {
         });
       }
 
-  RowMetadata getTotalsRowMetadata(ItemBase item) {
+  RowMetadata getTotalsRowMetadata(core.ItemBase item) {
     return new RowMetadata(
       selectable: false,
       focusable: totalsFocusable,
@@ -196,8 +211,8 @@ class RowMetadata {
   bool focusable;
   String cssClasses;
   Map<String,Column> columns = <String,Column>{};
-  Formatter formatter;
-  Editor editor;
+  fm.FormatterBase formatter;
+  ed.Editor editor;
 
   RowMetadata({this.selectable, this.focusable, this.cssClasses, this.columns, this.formatter, this.editor});
 }